rename: make-profile-screen 컴포넌트 create-profile-screen으로 변경 #22
