+++ src/screens/index.ts
@@ -2,6 +2,6 @@ import AuthScreen from './auth/AuthScreen';
 import SignInScreen from './auth/SignInScreen';
 import SignUpScreen from './auth/SignUpScreen';
 import EmailAuthenticationScreen from './auth/EmailAuthenticationScreen';
-import MakeProfileScreen from './auth/MakeProfileScreen';
+import CreateProfileScreen from './auth/CreateProfileScreen';
 
-export { AuthScreen, SignInScreen, SignUpScreen, EmailAuthenticationScreen, MakeProfileScreen };
+export { AuthScreen, SignInScreen, SignUpScreen, EmailAuthenticationScreen, CreateProfileScreen };
